Reproject rasters to shapefile CRS
If raster CRS didn't match EPSG:4326, the pro4 string of the raster
was changed to EPSG:4326 proj4 string. This caused intersected values
to be all NaN when working with non-EPSG:4326 data.

Instead of assuming EPSG:4326, the CRS is read from the target
shapefile. If raster CRS doesn't match, it will be reprojected to
shapefile CRS.